refactor(ui): standardize TUI command file naming
Remove `_command` suffix from TUI command files for consistent naming:
- export_command.py → export.py
- refresh_command.py → refresh.py
- help_command.py → help.py
- optimize_command.py → optimize.py

Also rename corresponding test files and update imports.
